git-tasks: handle enterprise url base path for access token installation
