+++ tasks/git/src/test/resources/wiremock/auth/__files/installation.json
@@ -5,26 +5,26 @@
     "login": "octocat",
     "id": 1,
     "node_id": "MDQ6VXNlcjE=",
-    "avatar_url": "https://github.com/images/error/octocat_happy.gif",
+    "avatar_url": "{{request.baseUrl}}/images/error/octocat_happy.gif",
     "gravatar_id": "",
-    "url": "https://api.github.com/users/octocat",
-    "html_url": "https://github.com/octocat",
-    "followers_url": "https://api.github.com/users/octocat/followers",
-    "following_url": "https://api.github.com/users/octocat/following{/other_user}",
-    "gists_url": "https://api.github.com/users/octocat/gists{/gist_id}",
-    "starred_url": "https://api.github.com/users/octocat/starred{/owner}{/repo}",
-    "subscriptions_url": "https://api.github.com/users/octocat/subscriptions",
-    "organizations_url": "https://api.github.com/users/octocat/orgs",
-    "repos_url": "https://api.github.com/users/octocat/repos",
-    "events_url": "https://api.github.com/users/octocat/events{/privacy}",
-    "received_events_url": "https://api.github.com/users/octocat/received_events",
+    "url": "{{request.baseUrl}}/api/v3/users/octocat",
+    "html_url": "{{request.baseUrl}}/octocat",
+    "followers_url": "{{request.baseUrl}}/api/v3/users/octocat/followers",
+    "following_url": "{{request.baseUrl}}/api/v3/users/octocat/following{/other_user}",
+    "gists_url": "{{request.baseUrl}}/api/v3/users/octocat/gists{/gist_id}",
+    "starred_url": "{{request.baseUrl}}/api/v3/users/octocat/starred{/owner}{/repo}",
+    "subscriptions_url": "{{request.baseUrl}}/api/v3/users/octocat/subscriptions",
+    "organizations_url": "{{request.baseUrl}}/api/v3/users/octocat/orgs",
+    "repos_url": "{{request.baseUrl}}/api/v3/users/octocat/repos",
+    "events_url": "{{request.baseUrl}}/api/v3/users/octocat/events{/privacy}",
+    "received_events_url": "{{request.baseUrl}}/api/v3/users/octocat/received_events",
     "type": "User",
     "site_admin": false
   },
   "repository_selection": "all",
-  "access_tokens_url": "{{request.baseUrl}}/app/installations/12345/access_tokens",
-  "repositories_url": "https://api.github.com/installation/repositories",
-  "html_url": "https://github.com/settings/installations/12345",
+  "access_tokens_url": "{{request.baseUrl}}/api/v3/app/installations/12345/access_tokens",
+  "repositories_url": "{{request.baseUrl}}/api/v3/installation/repositories",
+  "html_url": "{{request.baseUrl}}/settings/installations/12345",
   "app_id": 1430126,
   "app_slug": "mock-app-slug",
   "target_id": 54321,
